UltraBlame original commit: b2ac400bcb187648401def741146281224ed9bde

+++ js/src/jsemit.cpp
@@ -1858,7 +1858,7 @@ EmitEnterBlock(JSContext *cx, JSParseNode *pn, JSCodeGenerator *cg)
         JSDefinition *dn = (JSDefinition *) v.toPrivate();
         JS_ASSERT(dn->pn_defn);
         JS_ASSERT(uintN(dn->frameSlot() + depth) < JS_BIT(16));
-        dn->pn_cookie.set(dn->pn_cookie.level(), dn->frameSlot() + depth);
+        dn->pn_cookie.set(dn->pn_cookie.level(), uint16(dn->frameSlot() + depth));
 #ifdef DEBUG
         for (JSParseNode *pnu = dn->dn_uses; pnu; pnu = pnu->pn_link) {
             JS_ASSERT(pnu->pn_lexdef == dn);
@@ -1962,7 +1962,7 @@ MakeUpvarForEval(JSParseNode *pn, JSCodeGenerator *cg)
     }
 
     pn->pn_op = JSOP_GETUPVAR;
-    pn->pn_cookie.set(cg->staticLevel, ALE_INDEX(ale));
+    pn->pn_cookie.set(cg->staticLevel, uint16(ALE_INDEX(ale)));
     pn->pn_dflags |= PND_BOUND;
     return true;
 }
@@ -6731,8 +6731,7 @@ js_EmitTree(JSContext *cx, JSCodeGenerator *cg, JSParseNode *pn)
                 bool lambda = PN_OP(init) == JSOP_LAMBDA;
                 if (lambda)
                     ++methodInits;
-                if (op == JSOP_INITPROP && lambda && init->pn_funbox->joinable())
-                {
+                if (op == JSOP_INITPROP && lambda && init->pn_funbox->joinable()) {
                     op = JSOP_INITMETHOD;
                     pn2->pn_op = uint8(op);
                 } else {
